+++ spectator-ext-sandbox/src/main/java/com/netflix/spectator/sandbox/BucketFunction.java
@@ -15,10 +15,15 @@
  */
 package com.netflix.spectator.sandbox;
 
+import java.util.function.LongFunction;
+
 /**
  * Function to map an amount passed to a distribution summary or timer to a bucket.
+ *
+ * @deprecated Moved to {@code com.netflix.spectator.api.histogram} package. This is now just a
+ * thin wrapper to preserve compatibility. Scheduled for removal after in Q3 2016.
  */
-public interface BucketFunction {
+public interface BucketFunction extends LongFunction<String> {
   /**
    * Returns a bucket for the specified amount.
    *
