+++ spectator-ext-sandbox/src/main/java/com/netflix/spectator/sandbox/BucketFunctions.java
@@ -15,91 +15,22 @@
  */
 package com.netflix.spectator.sandbox;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.concurrent.TimeUnit;
+import java.util.function.LongFunction;
 
 /**
  * Helpers for creating bucketing functions.
+ *
+ * @deprecated Moved to {@code com.netflix.spectator.api.histogram} package. This is now just a
+ * thin wrapper to preserve compatibility. Scheduled for removal after in Q3 2016.
  */
 public final class BucketFunctions {
-
-  /**
-   * Predefined formatters used to create the bucket labels.
-   */
-  static final List<ValueFormatter> FORMATTERS = new ArrayList<>();
-
-  static {
-    FORMATTERS.add(fmt(TimeUnit.NANOSECONDS.toNanos(10),     1, "ns",  TimeUnit.NANOSECONDS));
-    FORMATTERS.add(fmt(TimeUnit.NANOSECONDS.toNanos(100),    2, "ns",  TimeUnit.NANOSECONDS));
-    FORMATTERS.add(fmt(TimeUnit.MICROSECONDS.toNanos(1),     3, "ns",  TimeUnit.NANOSECONDS));
-    FORMATTERS.add(fmt(TimeUnit.MICROSECONDS.toNanos(8),     4, "ns",  TimeUnit.NANOSECONDS));
-    FORMATTERS.add(fmt(TimeUnit.MICROSECONDS.toNanos(10),    1, "us",  TimeUnit.MICROSECONDS));
-    FORMATTERS.add(fmt(TimeUnit.MICROSECONDS.toNanos(100),   2, "us",  TimeUnit.MICROSECONDS));
-    FORMATTERS.add(fmt(TimeUnit.MILLISECONDS.toNanos(1),     3, "us",  TimeUnit.MICROSECONDS));
-    FORMATTERS.add(fmt(TimeUnit.MILLISECONDS.toNanos(8),     4, "us",  TimeUnit.MICROSECONDS));
-    FORMATTERS.add(fmt(TimeUnit.MILLISECONDS.toNanos(10),    1, "ms",  TimeUnit.MILLISECONDS));
-    FORMATTERS.add(fmt(TimeUnit.MILLISECONDS.toNanos(100),   2, "ms",  TimeUnit.MILLISECONDS));
-    FORMATTERS.add(fmt(TimeUnit.SECONDS.toNanos(1),          3, "ms",  TimeUnit.MILLISECONDS));
-    FORMATTERS.add(fmt(TimeUnit.SECONDS.toNanos(8),          4, "ms",  TimeUnit.MILLISECONDS));
-    FORMATTERS.add(fmt(TimeUnit.SECONDS.toNanos(10),         1, "s",   TimeUnit.SECONDS));
-    FORMATTERS.add(fmt(TimeUnit.SECONDS.toNanos(100),        2, "s",   TimeUnit.SECONDS));
-    FORMATTERS.add(fmt(TimeUnit.MINUTES.toNanos(8),          3, "s",   TimeUnit.SECONDS));
-    FORMATTERS.add(fmt(TimeUnit.MINUTES.toNanos(10),         1, "min", TimeUnit.MINUTES));
-    FORMATTERS.add(fmt(TimeUnit.MINUTES.toNanos(100),        2, "min", TimeUnit.MINUTES));
-    FORMATTERS.add(fmt(TimeUnit.HOURS.toNanos(8),            3, "min", TimeUnit.MINUTES));
-    FORMATTERS.add(fmt(TimeUnit.HOURS.toNanos(10),           1, "h",   TimeUnit.HOURS));
-    FORMATTERS.add(fmt(TimeUnit.HOURS.toNanos(100),          2, "h",   TimeUnit.HOURS));
-    FORMATTERS.add(fmt(TimeUnit.DAYS.toNanos(8),             1, "h",   TimeUnit.HOURS));
-    FORMATTERS.add(fmt(TimeUnit.DAYS.toNanos(10),            1, "d",   TimeUnit.DAYS));
-    FORMATTERS.add(fmt(TimeUnit.DAYS.toNanos(100),           2, "d",   TimeUnit.DAYS));
-    FORMATTERS.add(fmt(TimeUnit.DAYS.toNanos(1000),          3, "d",   TimeUnit.DAYS));
-    FORMATTERS.add(fmt(TimeUnit.DAYS.toNanos(10000),         4, "d",   TimeUnit.DAYS));
-    FORMATTERS.add(fmt(TimeUnit.DAYS.toNanos(100000),        5, "d",   TimeUnit.DAYS));
-    FORMATTERS.add(fmt(Long.MAX_VALUE,                       6, "d",   TimeUnit.DAYS));
-    // TimeUnit.NANOSECONDS.toDays(java.lang.Long.MAX_VALUE) == 106751
-  }
-
-  private static ValueFormatter fmt(long max, int width, String suffix, TimeUnit unit) {
-    return new ValueFormatter(max, width, suffix, unit);
-  }
 
   private BucketFunctions() {
   }
 
-  private static ValueFormatter getFormatter(long max) {
-    for (ValueFormatter f : FORMATTERS) {
-      if (max < f.max) {
-        return f;
-      }
-    }
-    return new ValueFormatter(max, ("" + max).length(), "ns", TimeUnit.NANOSECONDS);
-  }
-
-  private static BucketFunction timeBiasZero(String ltZero, String gtMax, long max, TimeUnit unit) {
-    final long nanos = unit.toNanos(max);
-    final ValueFormatter f = getFormatter(nanos);
-    final long v = f.unit().convert(max, unit);
-    List<Bucket> buckets = new ArrayList<>();
-    buckets.add(new Bucket(ltZero, 0L));
-    buckets.add(f.newBucket(v / 8));
-    buckets.add(f.newBucket(v / 4));
-    buckets.add(f.newBucket(v / 2));
-    buckets.add(f.newBucket(v));
-    return new ListBucketFunction(buckets, gtMax);
-  }
-
-  private static BucketFunction timeBiasMax(String ltZero, String gtMax, long max, TimeUnit unit) {
-    final long nanos = unit.toNanos(max);
-    final ValueFormatter f = getFormatter(nanos);
-    final long v = f.unit().convert(max, unit);
-    List<Bucket> buckets = new ArrayList<>();
-    buckets.add(new Bucket(ltZero, 0L));
-    buckets.add(f.newBucket(v - v / 2));
-    buckets.add(f.newBucket(v - v / 4));
-    buckets.add(f.newBucket(v - v / 8));
-    buckets.add(f.newBucket(v));
-    return new ListBucketFunction(buckets, gtMax);
+  private static BucketFunction wrap(LongFunction<String> f) {
+    return amount -> f.apply(amount);
   }
 
   /**
@@ -119,7 +50,7 @@ public final class BucketFunctions {
    *     so they can be used with a simple group by.
    */
   public static BucketFunction age(long max, TimeUnit unit) {
-    return timeBiasZero("future", "old", max, unit);
+    return wrap(com.netflix.spectator.api.histogram.BucketFunctions.age(max, unit));
   }
 
   /**
@@ -140,7 +71,7 @@ public final class BucketFunctions {
    *     so they can be used with a simple group by.
    */
   public static BucketFunction latency(long max, TimeUnit unit) {
-    return timeBiasZero("negative_latency", "slow", max, unit);
+    return wrap(com.netflix.spectator.api.histogram.BucketFunctions.latency(max, unit));
   }
 
   /**
@@ -160,7 +91,7 @@ public final class BucketFunctions {
    *     so they can be used with a simple group by.
    */
   public static BucketFunction ageBiasOld(long max, TimeUnit unit) {
-    return timeBiasMax("future", "old", max, unit);
+    return wrap(com.netflix.spectator.api.histogram.BucketFunctions.ageBiasOld(max, unit));
   }
 
   /**
@@ -181,95 +112,6 @@ public final class BucketFunctions {
    *     so they can be used with a simple group by.
    */
   public static BucketFunction latencyBiasSlow(long max, TimeUnit unit) {
-    return timeBiasMax("negative_latency", "slow", max, unit);
-  }
-
-  /**
-   * Format a value as a bucket label.
-   */
-  static class ValueFormatter {
-    private final long max;
-    private final String fmt;
-    private final TimeUnit unit;
-
-    /**
-     * Create a new instance.
-     *
-     * @param max
-     *     Maximum value intended to be passed into the apply method. Max value is in nanoseconds.
-     * @param width
-     *     Number of digits to use for the numeric part of the label.
-     * @param suffix
-     *     Unit suffix appended to the label.
-     * @param unit
-     *     Unit for the value in the label.
-     */
-    ValueFormatter(long max, int width, String suffix, TimeUnit unit) {
-      this.max = max;
-      this.fmt = "%0" + width + "d" + suffix;
-      this.unit = unit;
-    }
-
-    /** Return the max value intended for this formatter. */
-    long max() {
-      return max;
-    }
-
-    /** Return the unit for the formatter. */
-    TimeUnit unit() {
-      return unit;
-    }
-
-    /** Convert the value {@code v} into a bucket label string. */
-    String apply(long v) {
-      return String.format(fmt, unit.convert(v, TimeUnit.NANOSECONDS));
-    }
-
-    /** Return a new bucket for the specified value. */
-    Bucket newBucket(long v) {
-      final long nanos = unit.toNanos(v);
-      return new Bucket(apply(nanos), nanos);
-    }
-  }
-
-  private static class ListBucketFunction implements BucketFunction {
-    private final List<Bucket> buckets;
-    private final String fallback;
-
-    ListBucketFunction(List<Bucket> buckets, String fallback) {
-      this.buckets = buckets;
-      this.fallback = fallback;
-    }
-
-    @Override public String apply(long amount) {
-      for (Bucket b : buckets) {
-        if (amount < b.upperBoundary) {
-          return b.name();
-        }
-      }
-      return fallback;
-    }
-  }
-
-  private static class Bucket {
-    private final String name;
-    private final long upperBoundary;
-
-    Bucket(String name, long upperBoundary) {
-      this.name = name;
-      this.upperBoundary = upperBoundary;
-    }
-
-    String name() {
-      return name;
-    }
-
-    long upperBoundary() {
-      return upperBoundary;
-    }
-
-    @Override public String toString() {
-      return String.format("Bucket(%s,%d)", name, upperBoundary);
-    }
+    return wrap(com.netflix.spectator.api.histogram.BucketFunctions.latencyBiasSlow(max, unit));
   }
 }
